Assignment5 (#9) (#8)
* Added initial commit for assignment 5 (#1)

+++ webapp/spring-login-master/src/main/java/com/cloud/controller/TransactionController.java
@@ -7,8 +7,6 @@ import java.util.List;
 
 import javax.servlet.http.HttpServletResponse;
 
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -36,7 +34,7 @@ import com.cloud.util.Utils;
 @RestController
 public class TransactionController {
 
-	private static final Logger logger = LogManager.getLogger(TransactionController.class);
+	//private static final //logger //logger = LogManager.get//logger(TransactionController.class);
 	
 	@Autowired
 	private TransactionService transactionService;
@@ -59,7 +57,7 @@ public class TransactionController {
 	@ResponseBody
 	public TransactionWrapper findByUserId(HttpServletResponse response) throws IOException {
 		
-		logger.info("Find Transactions by User : Start");
+		//logger.info("Find Transactions by User : Start");
 		
 		TransactionWrapper transactions = new TransactionWrapper();
 		
@@ -72,12 +70,12 @@ public class TransactionController {
 			transactions.setStatusCode(CommonConstants.StatusCodes.SUCCESS);
 			transactions.setMessage(CommonConstants.SUCCESS);
 		} catch (Exception e) {
-			logger.error("Get all transactions for user failed");
+			//logger.error("Get all transactions for user failed");
 			transactions.setStatusCode(CommonConstants.StatusCodes.GET_ALL_TRANSACTIONS_FAILURE);
 			transactions.setMessage(CommonConstants.GET_ALL_TRANSACTION_FAILURE + ":" + e.getMessage());
 		}
 
-		logger.info("Find Transactions by User : End");
+		//logger.info("Find Transactions by User : End");
 		
 		return transactions;
 	}
@@ -92,7 +90,7 @@ public class TransactionController {
 	@ResponseBody
 	public Status create(@RequestBody Transaction transaction, HttpServletResponse response) throws IOException {
 
-		logger.info("Create Transaction - Start");
+		//logger.info("Create Transaction - Start");
 
 		Status status = new Status();
 
@@ -108,21 +106,21 @@ public class TransactionController {
 				SimpleDateFormat sf = new SimpleDateFormat(transaction.getDate().toString());
 				transaction.setDate(sf.format(new Date()));
 				transactionService.save(transaction);
-				status.setStatusCode(CommonConstants.StatusCodes.SUCCESS);
+				status.setStatusCode(CommonConstants.StatusCodes.TRANSACTION_SUCCESS);
 				status.setMessage(CommonConstants.SUCCESS);
 
 			} else {
-				logger.info("Unauthorized user");
+				//logger.info("Invalid Date format");
 				status.setStatusCode(CommonConstants.StatusCodes.INVALID_DATE_FORMAT);
 				status.setMessage(CommonConstants.INVALID_DATE_FORMAT);
 			}
 		} catch (Exception e) {
-			logger.error("Create transaction failed");
+			//logger.error("Create transaction failed");
 			status.setStatusCode(CommonConstants.StatusCodes.TRANSACTION_CREATION_FAILURE);
 			status.setMessage(CommonConstants.TRANSACTION_FAILURE + ":" + e.getMessage());
 		}
 
-		logger.info("Create Transaction - End");
+		//logger.info("Create Transaction - End");
 
 		return status;
 	}
@@ -138,7 +136,7 @@ public class TransactionController {
 	public Status update(@PathVariable String id, @RequestBody Transaction transaction, HttpServletResponse response)
 			throws IOException {
 		
-		logger.info("Update Transaction - Start");
+		//logger.info("Update Transaction - Start");
 		
 		Status status = new Status();
 		// Fetches the current user name who is logged in
@@ -152,20 +150,20 @@ public class TransactionController {
 				// Update the transaction with the new values
 				actualTransaction = this.setTransactionData(transaction, actualTransaction);
 				transactionService.save(actualTransaction);
-				status.setStatusCode(CommonConstants.StatusCodes.SUCCESS);
+				status.setStatusCode(CommonConstants.StatusCodes.TRANSACTION_SUCCESS);
 				status.setMessage(CommonConstants.SUCCESS);
 			} else {
-				logger.info("Unauthorized user");
+				//logger.info("Unauthorized user");
 				status.setStatusCode(CommonConstants.StatusCodes.UNAUTHORIZED);
 				status.setMessage(CommonConstants.UNAUTHORIZED);
 			}
 		} catch (Exception e) {
-			logger.error("Update transaction failed");
+			//logger.error("Update transaction failed");
 			status.setStatusCode(CommonConstants.StatusCodes.TRANSACTION_UPDATION_FAILURE);
 			status.setMessage(CommonConstants.TRANSACTION_FAILURE + " : " + e.getMessage());
 		}
 
-		logger.info("Update Transaction - End");
+		//logger.info("Update Transaction - End");
 		
 		return status;
 	}
@@ -181,7 +179,7 @@ public class TransactionController {
 	@ResponseBody
 	public Status delete(@PathVariable String id, HttpServletResponse response) throws IOException {
 		
-		logger.info("Delete Transaction - Start");
+		//logger.info("Delete Transaction - Start");
 		
 		Status status = new Status();
 		// Fetches the current user name who is logged in
@@ -192,20 +190,20 @@ public class TransactionController {
 
 			if (transaction.getUser().getEmail().equalsIgnoreCase(auth.getName())) {
 				transactionService.deleteById(id);
-				status.setStatusCode(CommonConstants.StatusCodes.SUCCESS);
+				status.setStatusCode(CommonConstants.StatusCodes.DELETION_SUCCESS);
 				status.setMessage(CommonConstants.SUCCESS);
 			} else {
-				logger.info("Unauthorized user");
+				//logger.info("Unauthorized user");
 				status.setStatusCode(CommonConstants.StatusCodes.UNAUTHORIZED);
 				status.setMessage(CommonConstants.UNAUTHORIZED);
 			}
 		} catch (Exception e) {
-			logger.error("Delete transactions failed");
+			//logger.error("Delete transactions failed");
 			status.setStatusCode(CommonConstants.StatusCodes.TRANSACTION_DELETION_FAILURE);
 			status.setMessage(CommonConstants.TRANSACTION_DELETION_FAILURE + ":" + e.getMessage());
 		}
 
-		logger.info("Delete Transaction - End");
+		//logger.info("Delete Transaction - End");
 		
 		return status;
 
@@ -218,32 +216,40 @@ public class TransactionController {
 	 * @throws IOException 
 	 */
 	@RequestMapping(value = "/transaction/{id}/attachments", method = RequestMethod.GET)
-    public AttachmentWrapper getReceipt(@PathVariable String id, HttpServletResponse response) throws IOException {
-		
-		logger.info("Get Transaction Receipt with id : "+ id + "Start");
-		
+	public AttachmentWrapper getReceipt(@PathVariable String id, HttpServletResponse response) throws IOException {
+
+		//logger.info("Get Transaction Receipt with id : " + id + "Start");
+
+		// Fetches the current user name who is logged in
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+
 		AttachmentWrapper attachmentWrapper = new AttachmentWrapper();
 		List<Attachment> attachmentList = null;
-		
-		//Fetches all the attachments in the transaction
+
+		// Fetches all the attachments in the transaction
 		try {
-			
 			Transaction transaction = transactionService.find(id);
-			attachmentList = transaction.getAttachments();
-			attachmentWrapper.setAttachments(attachmentList);
-			attachmentWrapper.setMessage(CommonConstants.SUCCESS);
-			attachmentWrapper.setStatusCode(CommonConstants.StatusCodes.SUCCESS);
-			
+			if (transaction.getUser().getEmail().equalsIgnoreCase(auth.getName())) {
+				attachmentList = transaction.getAttachments();
+				attachmentWrapper.setAttachments(attachmentList);
+				attachmentWrapper.setMessage(CommonConstants.SUCCESS);
+				attachmentWrapper.setStatusCode(CommonConstants.StatusCodes.SUCCESS);
+			} else {
+				//logger.info("Unauthorized user");
+				attachmentWrapper.setStatusCode(CommonConstants.StatusCodes.UNAUTHORIZED);
+				attachmentWrapper.setMessage(CommonConstants.UNAUTHORIZED);
+			}
+
 		} catch (Exception e) {
-			logger.error("Get transaction receipts failed");
+			//logger.error("Get transaction receipts failed");
 			attachmentWrapper.setStatusCode(CommonConstants.StatusCodes.GET_ATTACHMENT_FAILURE);
 			attachmentWrapper.setMessage(CommonConstants.GET_ATTACHMENTS_FAILURE + ":" + e.getMessage());
 		}
-		
-		logger.info("Get Transaction Receipt with id : "+ id + "- End");
-		
+
+		//logger.info("Get Transaction Receipt with id : " + id + "- End");
+
 		return attachmentWrapper;
-    }
+	}
 	
 	/**
 	 * Added to upload a receipt to a transaction
@@ -253,39 +259,45 @@ public class TransactionController {
 	@RequestMapping(value = "/transaction/{id}/attachments", method = RequestMethod.POST)
 	public Status uploadReceipt(@PathVariable String id, @RequestPart(value = "file") MultipartFile file) {
 
-		logger.info("Attach Transaction Receipt with id : " + id + " - Start");
-		
+		//logger.info("Attach Transaction Receipt with id : " + id + " - Start");
+
+		// Fetches the current user name who is logged in
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+
 		Status status = new Status();
 
-		try 
-		{
-			if(Utils.isValidExt(file))
-			{
-				// Upload the receipt
-				String uri = baseClient.uploadFile(file);
+		try {
+			if (Utils.isValidExt(file)) {
+				Transaction transaction = transactionService.find(id);
+				if (transaction.getUser().getEmail().equalsIgnoreCase(auth.getName())) {
+					// Upload the receipt
+					String uri = baseClient.uploadFile(file);
 
-				// Save the metadata of the receipt in the database attachment table
-				transactionService.saveAttachment(id, uri);
-				status.setMessage(uri);
-				status.setStatusCode(CommonConstants.StatusCodes.SUCCESS);
+					// Save the metadata of the receipt in the database attachment table
+					transactionService.saveAttachment(id, uri);
+					status.setMessage(uri);
+					status.setStatusCode(CommonConstants.StatusCodes.SUCCESS);
+				} else {
+					//logger.info("Unauthorized user");
+					status.setStatusCode(CommonConstants.StatusCodes.UNAUTHORIZED);
+					status.setMessage(CommonConstants.UNAUTHORIZED);
+				}
 
-			}
-			else
-			{
+			} else {
 				status.setStatusCode(CommonConstants.StatusCodes.INVALID_ATTACHMENT);
 				status.setMessage(CommonConstants.INVALID_ATTACHMENT);
-				logger.error("Invlaid file extension");
+				//logger.error("Invlaid file extension");
 			}
-						
+
 		} catch (Exception e) {
-			
+
 			status.setStatusCode(CommonConstants.StatusCodes.INVALID_ATTACHMENT);
 			status.setMessage(CommonConstants.UPLOAD_ATTACHMENTS_FAILURE + e.getMessage());
-			logger.error("Error while attaching the receipt");
+			//logger.error("Error while attaching the receipt");
 		}
 
-		logger.info("Attach Transaction Receipt with id : " + id + " - End");
-		
+		//logger.info("Attach Transaction Receipt with id : " + id + " - End");
+
 		return status;
 	}
 	
@@ -295,58 +307,68 @@ public class TransactionController {
 	 * @return
 	 */
 	@RequestMapping(value = "/transaction/{id}/attachments/{attachmentId}", method = RequestMethod.PUT)
-	public Status updateReceipt(@PathVariable String id, @PathVariable String attachmentId, @RequestPart(value = "file") MultipartFile file) {
+	public Status updateReceipt(@PathVariable String id, @PathVariable String attachmentId,
+			@RequestPart(value = "file") MultipartFile file) {
 
-		logger.info("Attach Transaction Receipt with id : " + id + " - Start");
-		
+		//logger.info("Attach Transaction Receipt with id : " + id + " - Start");
+
+		// Fetches the current user name who is logged in
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+
 		Status status = new Status();
 		boolean receiptPresent = false;
 
-		try 
-		{
+		try {
 			if (Utils.isValidExt(file)) {
 				// Check if attachment is present
 				Transaction transaction = transactionService.find(id);
-				Attachment oldAttachment = null;
-				for (Attachment attachment : transaction.getAttachments()) {
-					if (attachment.getId().toString().equals(attachmentId)) {
-						receiptPresent = true;
-						oldAttachment = attachment;
-						break;
+				if (transaction.getUser().getEmail().equalsIgnoreCase(auth.getName())) {
+
+					Attachment oldAttachment = null;
+					for (Attachment attachment : transaction.getAttachments()) {
+						if (attachment.getId().toString().equals(attachmentId)) {
+							receiptPresent = true;
+							oldAttachment = attachment;
+							break;
+						}
 					}
-				}
 
-				if (receiptPresent) {
-					// Delete the existing file
-					baseClient.deleteFile(oldAttachment.getUri());
-					// Upload the receipt
-					String uri = baseClient.uploadFile(file);
-					oldAttachment.setUri(uri);
-					// Save the metadata of the receipt in the database attachment table
-					transactionService.save(oldAttachment);
-					status.setMessage(uri);
-					status.setStatusCode(CommonConstants.StatusCodes.SUCCESS);
+					if (receiptPresent) {
+						// Delete the existing file
+						baseClient.deleteFile(oldAttachment.getUri());
+						// Upload the receipt
+						String uri = baseClient.uploadFile(file);
+						oldAttachment.setUri(uri);
+						// Save the metadata of the receipt in the database attachment table
+						transactionService.save(oldAttachment);
+						status.setMessage(uri);
+						status.setStatusCode(CommonConstants.StatusCodes.SUCCESS);
 
+					} else {
+						//logger.info("Receipt not present for the transaction");
+						status.setMessage(CommonConstants.ATTACHMENTS_NOT_PRESENT);
+						status.setStatusCode(CommonConstants.StatusCodes.ATTACHMENT_NOT_PRESENT);
+					}
 				} else {
-					logger.info("Receipt not present for the transaction");
-					status.setMessage(CommonConstants.ATTACHMENTS_NOT_PRESENT);
-					status.setStatusCode(CommonConstants.StatusCodes.ATTACHMENT_NOT_PRESENT);
+					//logger.info("Unauthorized user");
+					status.setStatusCode(CommonConstants.StatusCodes.UNAUTHORIZED);
+					status.setMessage(CommonConstants.UNAUTHORIZED);
 				}
 			} else {
 				status.setStatusCode(CommonConstants.StatusCodes.INVALID_ATTACHMENT);
 				status.setMessage(CommonConstants.INVALID_ATTACHMENT);
-				logger.error("Invlaid file extension");
+				//logger.error("Invlaid file extension");
 			}
 
 		} catch (Exception e) {
-			
+
 			status.setStatusCode(CommonConstants.StatusCodes.UPLOAD_ATTACHMENT_FAILURE);
 			status.setMessage(CommonConstants.UPLOAD_ATTACHMENTS_FAILURE + e.getMessage());
-			logger.error("Error while attaching the receipt");
+			//logger.error("Error while attaching the receipt");
 		}
 
-		logger.info("Attach Transaction Receipt with id : " + id + " - End");
-		
+		//logger.info("Attach Transaction Receipt with id : " + id + " - End");
+
 		return status;
 	}
 
@@ -360,46 +382,53 @@ public class TransactionController {
 	public Status deleteAttachment(@PathVariable String id, @PathVariable String attachmentId,
 			HttpServletResponse response) throws IOException {
 
-		logger.info("Delete Transaction Receipt with id : " + id + "- Start");
-		
+		//logger.info("Delete Transaction Receipt with id : " + id + "- Start");
+
+		// Fetches the current user name who is logged in
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+
 		Status status = new Status();
 		boolean receiptPresent = false;
 
 		// Save the metadata of the receipt in the database attachment table
 		try {
-			
+
 			Transaction transaction = transactionService.find(id);
-			Attachment oldAttachment = null;
-			for(Attachment attachment : transaction.getAttachments())
-			{
-				if(attachment.getId().toString().equals(attachmentId))
-				{
-					receiptPresent = true;
-					oldAttachment = attachment;
-					break;
+			if (transaction.getUser().getEmail().equalsIgnoreCase(auth.getName())) {
+				Attachment oldAttachment = null;
+				for (Attachment attachment : transaction.getAttachments()) {
+					if (attachment.getId().toString().equals(attachmentId)) {
+						receiptPresent = true;
+						oldAttachment = attachment;
+						break;
+					}
 				}
-			}
-			
-			if (receiptPresent) {
-				String result = baseClient.deleteFile(oldAttachment.getUri());
-				transactionService.deleteAttachment(id, oldAttachment.getUri());
-				status.setMessage(result);
-				status.setStatusCode(CommonConstants.StatusCodes.SUCCESS);
+
+				if (receiptPresent) {
+					String result = baseClient.deleteFile(oldAttachment.getUri());
+					transactionService.deleteAttachment(id, oldAttachment.getUri());
+					status.setMessage(result);
+					status.setStatusCode(CommonConstants.StatusCodes.DELETION_SUCCESS);
+				} else {
+					//logger.info("Receipt not present for the transaction");
+					status.setMessage(CommonConstants.ATTACHMENTS_NOT_PRESENT);
+					status.setStatusCode(CommonConstants.StatusCodes.ATTACHMENT_NOT_PRESENT);
+				}
 			} else {
-				logger.info("Receipt not present for the transaction");
-				status.setMessage(CommonConstants.ATTACHMENTS_NOT_PRESENT);
-				status.setStatusCode(CommonConstants.StatusCodes.ATTACHMENT_NOT_PRESENT);
+				//logger.info("Unauthorized user");
+				status.setStatusCode(CommonConstants.StatusCodes.UNAUTHORIZED);
+				status.setMessage(CommonConstants.UNAUTHORIZED);
 			}
-			
+
 		} catch (Exception e) {
-			
+
 			status.setStatusCode(CommonConstants.StatusCodes.ATTACHMENT_DELETION_FAILURE);
 			status.setMessage(CommonConstants.DELETE_ATTACHMENTS_FAILURE + e.getMessage());
-			logger.error("Error while deleting a receipt");
+			//logger.error("Error while deleting a receipt");
 		}
-		
-		logger.info("Delete Transaction Receipt with id : " + id + "- End");
-		
+
+		//logger.info("Delete Transaction Receipt with id : " + id + "- End");
+
 		return status;
 
 	}
@@ -412,7 +441,7 @@ public class TransactionController {
 	 */
 	private Transaction setTransactionData(Transaction transaction, Transaction actualTransaction) {
 
-		logger.info("Set Transaction Data - Start");
+		//logger.info("Set Transaction Data - Start");
 		
 		actualTransaction.setAmount(transaction.getAmount());
 		actualTransaction.setDate(transaction.getDate());
@@ -420,7 +449,7 @@ public class TransactionController {
 		actualTransaction.setDescription(transaction.getDescription());
 		actualTransaction.setMerchant(transaction.getMerchant());
 
-		logger.info("Set Transaction Data - End");
+		//logger.info("Set Transaction Data - End");
 		
 		return actualTransaction;
 	}
